Went a completely different direction for fixing #5: the issue stemmed from a conflicting package name, between the :std tag library created in the generated skeleton project...and a nickname for the ```swank-trace-dialog``` package.
Since the real purpose of the :std was more or less to demonstrate how to create a "standard" library of tags that define the basic components of a site's pages, I've renamed the :std tag library to be :ui-theme--that is closer to its real purpose, after all.

This also required fixing up the :docs tag library that ships with :uh-web, as it assumes that a :std (now :ui-theme) tag library exists: that way, the pages it renders match the site into which docs are loaded.

Meanwhile, the related issue in slime may stay open or unresolved; at least this way, hh-web users need take no additional workaround.